Replace Mocha with Jest
Since Jest will be needed for snapshot testing.
