+++ package.json
@@ -13,12 +13,11 @@
 	"scripts": {
 		"deploy-site": "npm run catalog-build && storybook-to-ghpages --ci --existing-output-dir=catalog/build",
 		"eslint": "eslint ./components ./catalog --ext .js --ext .jsx",
-		"test": "npm run build && npm run unit-test && npm run eslint && npm run stylelint",
+		"test": "npm run build && jest && npm run eslint && npm run stylelint",
 		"precommit": "prettier --write \"./components/**/*.{js,jsx}\" && npm run stylelint",
 		"build": "rimraf dist && webpack --mode=none",
 		"prepublishOnly": "npm test",
 		"stylelint": "stylelint ./components/**/*.jsx",
-		"unit-test": "cross-env NODE_ENV=test mocha -t 60000 -r @babel/register ./tests/**/*.js",
 		"ship-to": "npx --userconfig=./.npmrc @faithlife/helmsman ship-to @faithlife/styled-ui",
 		"tag-for-beta": "npx --userconfig=./.npmrc @faithlife/helmsman ship-to @faithlife/styled-ui beta",
 		"tag-for-production": "npx --userconfig=./.npmrc @faithlife/helmsman ship-to @faithlife/styled-ui production",
@@ -90,6 +89,7 @@
 		"autoprefixer": "^8.2.0",
 		"babel-core": "^7.0.0-bridge.0",
 		"babel-eslint": "^10.0.2",
+		"babel-jest": "^26.6.3",
 		"babel-loader": "^8.0.6",
 		"babel-plugin-add-module-exports": "^1.0.2",
 		"babel-plugin-react-docgen": "^3.1.0",
@@ -105,9 +105,9 @@
 		"eslint-config-faithlife": "^6.0.0",
 		"esm": "^3.0.24",
 		"image-webpack-loader": "^4.3.1",
+		"jest": "^26.6.3",
 		"micro-graphql-react": "^0.3.0-beta2",
 		"mini-css-extract-plugin": "^0.4.1",
-		"mocha": "^5.1.1",
 		"node-sass": "^4.14.1",
 		"postcss-inline-svg": "^3.1.1",
 		"postcss-loader": "^2.1.6",
